Save System, AK-47 and misc.
Implemented a save system and gave a tutorial in the Discord channel. Still needs some work on it.
Added an AK-47 but doesn't work properly right now.
Started on a Toggleable abstract class.
Many changes to the Generator script to be more generic and readable. Updated effected scripts as well.